requirements.txt
python-docx==0.8.11
tqdm==4.65.0